Sport App With Flutter
My graduation project with Flutter. To put it simply, there are options to register to the application with Gmail or log in directly with Google. If you register with Gmail, you can also use Gmail to log in. I forgot my password and account verification is also available. Users who log in for the first time are allowed to start from the pages where they fill in their user information (for example, height, weight, age). Afterwards, the user is welcomed on the home page and the application contains sports movements, recipes and sports articles. Firebase was used as the database. All articles etc. can be added via the database. Sports movements can also be sorted according to difficulty.